Fix temporal alarms - use longs in intents for stopcodes
git-svn-id: http://redbus.googlecode.com/svn/trunk@417 047a1783-e752-e205-4d80-9245024c4953

+++ src/org/redbus/ui/CreateShortcutActivity.java
@@ -59,7 +59,7 @@ public class CreateShortcutActivity extends ListActivity {
     	shortcutIntent.setClass(this, ArrivalTimeActivity.class);
         shortcutIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
         shortcutIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-        shortcutIntent.putExtra("StopCode", stopCode);
+        shortcutIntent.putExtra("StopCode", (long) stopCode);
 
         // Then, set up the container intent (the response to the launcher)
         Intent intent = new Intent();
